flask_catalog/requirements.txt
flask
pymysql
flask-sqlalchemy
flask-script
flask-wtf
flask-migrate
py-bcrypt
python-slugify
flask-markdown
Flask-Uploads==0.2.0
oauth2
urllib


